Improved DVBv5 API support
Added detection of DVBv5 and DVBv3 capabilities similar to how this is done in libdvbv5.
When DVBv5 capability is present use only DVBv5 API calls; otherwise, use only DVBv3 API calls.
Previously, for getting the signal strength etc first the DVBv3 call was done and when that
failed the DVBv5 call was done. This caused problems with the dvbloopback virtual driver.
Doing it the other way around, first the DVBv5 call and when that fails do the DVBv3 call,
caused kernel crashes on older systems that only supported DVBv3.
Previously, for tuning of DVB-T/T2/S/S2 the DVBv5 API was always used and for DVB-C and ATSC
the DVBv3 API was always used.  There have been no problems reported for this but it is now
done accordingly to the detected API version.

Refs trac-12638