Increase UDP buffer size for Sat>IP
Try to increase the UDP buffer size for reading the Sat>IP RTP streams to 8Mbytes (decimal).
A log message is given when the system maximum only allows for a lower value.
Change log messages to have only the cardid and not the device name on each message.
This leaves more space for the actual message text.

Refs #trac-13121