Convert preview generation to std::chrono. (33)
This required some rework of the functions called. Previously they
passed around a 'long long' that was either a time offset or a frame
offset, and passed a second variable to indicate the type of the first
variable. This has been changed to pass a std::chrono::seconds time
offset and a 'long long' frame offset. The time offset has priority,
so if it is > 0 and < max() then it will be used, otherwise the frame
offset is used. This should mimic the previous behavior.