Wayland: Remove Qt 'QT_WAYLAND_DISABLE_WINDOWDECORATION' workaround
- setting QT_WAYLAND_DISABLE_WINDOWDECORATION=1 was previously used as a
workaround to ensure that Qt did not force the alpha depth for the
default framebuffer to 0 - thus breaking alpha blending in the UI code.
- this is now left to the user to set if wanted/required
- with the 'waylandextras' support in master (if compiled in), the
workaround should not be required for sane compositors (weston, sway)
- without 'waylandextras' support, the setting does not work as expected
in more recent versions of Qt - which for some reason drop down to a
lower color depth (e.g. 556 instead of 888) with
QT_WAYLAND_DISABLE_WINDOWDECORATION=1
- so there are visual issues with or without the override - hence leave
it to the user to decide...