libmythtv: Store supported render formats in MythVideoFrame
- primarily designed to avoid multiple round trip calls from the decoder
through to the player and video output classes - across threads.
- slightly complicated by the need to have a copy of the render formats
in the decoder class to handle the av_get_buffer callback - which only
retrieves the next video frame if it needs to (i.e. when it is using
direct rendering).
- Note: the list of available direct render formats should never change
during the lifetime of the decoder, player or video output instances.