Channelscan dialog default choice for popup menu
In mythtv-setup there can be duplicate channel numbers found after a scan.
It is possible to resolve this by responding to a popup menu message with
a question for each conflicting channel.
If this question is asked for a number of channels the default choice
is now the answer that has been given for the previous channel.
Previously, the default choice was always the first menu entry.