Minimum time for acquiring DVB tables
The minimum time for acquiring tables is now set to 1000 milliseconds.
This is the difference between the Tuning Timeout and the Signal Timeout
as configured for the capture card.
According to the comment this should be at least 100 milliseconds but if the
value was less it was set to 2500 milliseconds.
The implementation is now to set it to 1000 milliseconds if the configured
value is less than 1000 milliseconds.