EIT event flow rate control
Processing EIT events is limited to 20 events per 400 milliseconds per thread to insure
that the time-critical functions of the backend are only minimally influenced.
Minor rate fluctuations are handled with an event queue for the incoming events.
Flow rate control is now implemented by discarding events when the event queue is full.
Previously, flow rate control was implemented by adding and removing the EIT pids
from the currently active PID filter depending on the number of events in the queue.
This was not particularly elegant as it lead to continuous on/off switching in some
situations and each change caused error messages about discontinuities in the packet stream.
Note that processing transport stream packets and creating EIT events is relatively
simple but processing EIT events including updating the database is much more work.