Search XMLTV grabbers only once in mythtv-setup
When the Video Sources page is selected then it has a popup
list with all XMLTV grabbers available on the system.
This popup list is the same for all video sources.
Previously, the search for the list of grabbers was executed for
each video source each time the Video Sources page was selected.
The search is now done only once and the search resuls are re-used.