Recording profile name readonly instead of disabled
Set the Profile Name field for recording profiles Default and Live TV
readonly instead of disabled. This keeps the field label "Profile Name"
visible instead of greyed-out. This also makes keyboard navigation in
the profile editing page of profiles with non-editable names, Default
and Live TV, consistent with that of the other profiles of whom the name
can be edited.