Commandline option --dvbv3 to force legacy DVBv3 API
Add commandline option --dvbv3 to force mythbackend to use the legacy DVBv3 API calls instead
of the DVBv5 API calls to retrieve signal strength, signal-to-noise ratio etc from DVB cards.
Other features, such as the setting of the modulation system which is essential for cards
supporting more than one modulation system, are still implemented with the DVBv5 API.
This feature is added to improve interoperability of MythTV with older and/or
badly supported DVB card drivers.

Refs #trac-12638