Handle sending UDP when using wildcard listens.
This problem was noticed while diagnosing a problem with AirPlay. The
code was unable to send any UDP packets back to the airplay server,
because all of attached PrivUdpSockets had an invalid address. Tweak
the code to not only try and find a socket with an appropriate network
range, but if that fails to fall back to looking for a wildcard socket
of the appropriate protocol.