libmythtv: Initial support for DRM video rendering
- *HIGHLY EXPERIMENTAL* and *VERY BUGGY*
- to enable runtime support, set the environment variable
MYTHTV_DRM_VIDEO and ensure the Qt QPA platform is explicitly set to
eglfs (either via the -platform command line argument or the
QT_QPA_PLATFORM environment variable)
- I recommend setting 'QT_LOGGING_RULES=qt.qpa.*=true' as well to confirm
that the MythTV internals are consistent with the Qt view of the DRM
world.
- there is a lot going on here - and some of it even works.
- on all platforms tested (amdgpu, pi4 and tanix tx6), video is
displayed when supplied via a DRM decoder (v4l2_m2m, v4l2_request and
VAAPI with DRM).
- that's about as far as the good news gets. On all platforms there is
significant tearing and/or ordering issues with the video, sundry alpha
blending issues with the video and GUI planes and a number of
predictable, unrecoverable Qt rendering failures with AMD.
- Note: I'm fairly sure that no other player supports DRM rendering of
VAAPI decoded frames - and it is only included here for testing
purposes. If it does not work well on Intel based systems it will be
removed.