Update Airplay to handle NTP timestamps properly.
NTP timestamps are transmitted as seconds and fractional seconds with
a epoch start of 1900-01-01. Make sure to add/subtract the conversion
from the Unix epoch to the NTP epoch, and properly convert the
fractional seconds field.

The existing code was copying timestamp sec/usec values to/from the
two halves of the NTP timestamp, which should not have caused
functional problems as the mistakes on sending and receiving a
reflected value would cancel each other.

This would have affected the computation of the clock skew value, but
since that value is never used there is no harm.