Initial commit of new HTTP and Websocket code
Note: the following functionality is incomplete and/or missing:-
 - authentication (easy enough)
 - WSDL (started - but will take time a little effort to complete)
 - actual handling of websocket functionality (i.e. remote procedure
calls)
 - no existing service APIs are converted in this commit - and each will
be require an update
 - no conversion of UPnP end points

- the principle aims here are to remove the dependency on QScript (which
is end of life), invert the dependency between HTTP and UPnP (i.e. UPnP
should require HTTP and not the other way around), extend functionality,
improve performance and memory footprint and ease maintenance.
- this is entirely standalone and can be run alongside the existing
server implementation (port settings are currently ignored and the
server will listen on 8081 and 8091 (SSL))
- the intention is that the new code behaves in exactly the same way as
the old code - with the exception of extending functionality.
- websocket handling is currently embedded and is fully compliant with
the autobahn test suite
- the server code is entirely lock free, event driven and thread safe -
but there will be instances where services require locking to process
requests