EIT Lookup channel ID from database
Add sanity check on video source value; do not search for
the chanid in the database if the video source is not known.
For DVB lookup select the channels also on video source instead
of later checking the video source on all channels found.
Removed the "fuzzy matching" when no match is found.
The "fuzzy matching" did not match on transport ID and could
thus return an incorrect channel ID value.