Error message for missing database setting
When database table settings does not have a value for BackendServerAddr
for the current host then mythbackend fails with an error message.
This happens when the database is created with the mc.sql file
but mythtv-setup has never been used to examine and possibly modify
the configuration values.
This is now expressed in the error message.
Previously, the error message referred to the setting of BackendServerIP
but that has been changed to BackendServerName with update to
database schema version 1346 in commit e6f11ce287d467053de719402fbcad5f6fb8047c