Crash in libCEC on mythfrontend GUI rebuild
Fix mythfrontend crash in GUI rebuild due to a theme change or due
to a change in the window size. This happens only when mythfrontend
is used with a monitor or a TV that does not support CEC.
In that case the CEC library is closed after startup but due
to a bug in libcec/cecloader.h the handle to libcec.so is not cleared.
On theme change or window resize the GUI is rebuilt and the CEC
library is opened a second time. This uses the now-invalid handle
and this causes the crash.
The real solution is to fix libcec.
For now the problem is solved by not closing the CEC library so that
the next time the handle is still valid.
See MythTV issue #299 and Pulse-Eight/libcec issue #555 on Github.

Fixes #299