eitfixup: A couple of tweaks to the Dutch fixups.
1) Use the "person separator" regex for actors. This allows the use of
the word "en" (and) between names as well as a simple comma.

2) Fix the code to extract the director's name.

3) Remove positive look ahead from two regular expressions. As written
these two expressions can never match as they require a space and a
non-space character to occupy the same position in the string. Without
positive look ahead they can match the words "uit"/"van" and the
following text.