Add FreeSync and GSync support (Variable Refresh Rate/VRR)
- Note: FreeSync and GSync seem to work fine with mythfrontend; these
changes allow certain optimisations when VRR is detected and is enabled
and under some situations allow the user to force VRR on or off for
mythfrontend (and mythavtest).
- GSync support is linux only when using Display Port, X11 and the
proprietary NVidia driver. It can be enabled/disabled via the NVidia
control/settings thingy or it can be forced on or off for mythtv with the
-/--vrr command line option (0 is off, >=1 on) - otherwise it will detect
the correct state and act accordingly.
- FreeSync support is detectable with AMD on linux - and works with
Display Port when using X11 (there is meant to be some wayland support
with sway but this has not been confirmed). It can be enabled via
xorg.conf but is otherwise not configurable for mythfrontend at run time.
If running mythfrontend without X11 or Wayland, it can again be forced
on or off with the -/--vrr (0 or 1) command line option; but for some
reason does not seem to work corrrectly (needs further investigation;
maybe swap interval/vsync related).