Crash in libCEC on mythfrontend GUI rebuild (#2)
Fix mythfrontend crash in GUI rebuild due to a theme change or due
to a change in the window size. This happens only when mythfrontend
is used with a monitor or a TV that does not support CEC.
In that case the CEC library is closed after startup but due
to a bug in libcec/cecloader.h the handle to libcec.so is not cleared.
On theme change or window resize the GUI is rebuilt and the CEC
library is opened a second time. This uses the now-invalid handle
and this causes the crash.
The real solution is to fix libcec.
For now the problem is solved by zeroing the handle ourselves, but this
requires accessing the handle which should be internal to the CEC library.
See also Pulse-Eight/libcec issue #555 on Github.
This fix replaces the first fix in which the library was never closed.
See commit 10b5291b832bc89ce726c5b53ac87ef738c2c3cd

Refs #299