Restore channel data from deleted channels
The channel table created by a mythtv-setup channel scan is filled
with data received from the broadcaster such as channel name and number.
There are additional fields in the channel table for additional data such
as the xmltvid, the icon filename and the visibility status of the channel.
The values of these fields are preserved when the channel is updated.
When channels have been deleted and are found again in a new
channel scan then these fields are all empty.
Since commit 2e26f7f12fce13c155e43ff536b1e09a342e1722,
"Preserve data for deleted channels that are still referenced.",
a channel that is deleted is not immediately really deleted
but it is marked for deletion later, so it does still exist for a while.
This makes it possible to retrieve the non-scanned data for
a new channel from the deleted channel when it is the same channel.
Note that a channel is considered to be the same when the network ID,
the transport ID and the service ID are equal.
The icon filename and the xmltvid are searched for in
all video sources, also including channels that are not deleted.
The visibility status is only retrieved from deleted channels
in the same video source because it is possible to have the same
channel in two video sources, one channel visible and one not.
The new "Restore Data" button is found in the Channel Editor of mythtv-setup.
This button is now defined in the MythCenter-wide theme but
not yet in other themes.