Channel Editor button Restore Data in default themes
In the mythtv-setup Channel Editor the Restore Data button is
added to the themes default and default-wide.
The theme default is part of the MythCenter theme.
When a theme is selected that does not have the Restore Data
button defined then this functionality is not available but
all other functionality is still present.
See also commit e0cf8c476cf171fba6e1dfb40b4bac28f7a99ad9
where the Restore Data functionality is introduced.