No multirec for V4L2ENC and HDPVR devices
In commit f58f474bb1 on 6 nov. 2019 the default for field schedgroup in new entries in
table capturecard is set to 1 for all types of devices.
This is the "Schedule as Group" setting that controls, among others, multirec.
This means that since that commit multirec is by default enabled for all new capturecard entries.
Multirec means that multiple recordings can be made simultaneously from one capturecard;
the term multirec in MythTV refers both to recording different channels from one multiplex
and to overlapping recordings of the same channel.

The devices that encode an analog signal, the V4L2ENC and the HDPVR, can obviously receive only
one channel at a time. They can however in principle support overlapping recordings but
in MythTV configuration there is no differentiation between this form of multirec and the multirec
that allows recording different channels.

For both the V4L2ENC and the HDPVR the default for field schedgroup in new entries in
table capturecard is now set to 0, disabling multirec for these types of devices.
The general default, as defined in the database schema, is not changed.

Note that the V4L2ENC devices have partial support for multirec and hence for these devices can,
in the "Input Connections" dialog of mythtv-setup, the "Schedule as Group" be checked or unchecked.
Multirec recordings will however fail.

The HDPVR devices have no support for multirec at all and therefore they do not have
the "Schedule as Group" configuration option.

See also ticket #trac-13501.

(cherry picked from commit 40e0a9889442c0828a499e34efc9bef137648097)
Signed-off-by: Klaas de Waal <kdewaal@mythtv.org>