Show "Probed Info" readonly instead of disabled
In the mythtv-setup Capture Card editor, show the "Probed Info" of HDPVR devices in a readonly
field instead of in a disabled field. A disabled field is greyed out and cannot be navigated. Previously,
before the readonly status was available, the disabled status was used to make the field readonly.