No multirec for MPEG, MJPEG, GO7007 and V4L analog input cards
On creation of capture card entries for analog input cards of type MPEG, MJPEG, GO7007 and V4L
set the database value schedgroup to 0. This disables multirec, similar to how that is
now done for the V4L2ENC and HDPVR devices.
For these cards there is no "Schedule as Group" field in the configuration dialog.
See also:
commit f58f474bb1df2cd8fdd14c81cceb02a9984f928e Fix issues with default schedgroup and reclimit values
commit 40e0a9889442c0828a499e34efc9bef137648097 No multirec for V4L2ENC and HDPVR devices