qt6: Convert old style power recording rule checks to QRegularExpression.
Apparently in the past power recording rules needed to start with
"AND", but is no longer necessary or allowed. There are several places
in the code that check for a leading "AND" in a SQL fragment, and
there are several places that just strip this keyword and continue
processing. Convert all of these functions to use a QRegularExpression
to match on the word "AND".