CEA-608/VBI CC3 closed captions/subtitles
The CEA-608 closed captions can show two different subtitle streams
for two different languages, called CC1 and CC3.
When present, these streams can be selected in the subtitle menu.
Due to a bug only the CC1 stream is actually shown.
This is now fixed and also the CC3 stream can now be shown.

Refs #326