Favor artwork from official source
The TvMaze website sets a flag called Main to True when the source of the image is the network which produced the show. Usually these are high quality images, so we want to default to these "official" images. This is done by putting them first in the list. By default, the first item of each artwork type (Coverart, Fanart, Banner) is the one which gets automatically assigned.
In addition, after consulting the MythTV_Universal_Metadata_Format page, my impression is that during the query operation, ./grabbername -M "title", one one of each artwork type is expected, though this is not explicitly specified. It's during the collection operation, ./grabbername -C <tvmaze_id>, that we need ALL of the artworks listed. I modified the script to follow this convention and then tested both automatic metadata assignment, and manual override (e.g. Metadata Options -> Find Coverart Online) to validate that all of the artwork options are presented to the user.