Adding output of popularity
The code which handles automatic assignment of metadata uses "popularity" as one of the factors when deciding between multiple database entries with exact title matches. In the TvMaze scheme popularity shows up as the 'weight' field. This commit change adds output of the popularity, and also adds output of userrating to the '-M' and '-C' versions. It was already in the '-D' version.