Fix ProgramInfo::ToMap to not lose data.
There are nine fields in this map whose values are presented as a
single UTF16 character, two of which are likely already overflowing
that value. To prevent changes to existing usage, new fields were
added to present the values in a different way. All nine have had a
"<id>_str" field added to the map that presents the number as a string
so there is no possibility of overflow.  I.E. The program flags might
be presented as the four character string "1538" instead of a one
character string containing QChar(1538).  Four of these values are
bitfields, and have also had an "<id>_names" field added to the map
that lists the textual names of the
bits. I.E. "CUTLIST|WATCHED|PRESERVED"

Fixes #337