Fix RecordingRule::ToMap to not lose data.
There is one field in this map whose value is presented as a single
UTF16 character, and is likely already overflowing that value. To
prevent changes to existing usage, a new field were added to present
the values in a different way. It has had a "<id>_str" field added to
the map that presents the number as a string so there is no
possibility of overflow.  I.E. The chanid field might be presented as
the four character string "1538" instead of a one character string
containing QChar(1538).

Fixes #342