qt6: Rework the widget focus code for QMultiMap changes.
In Qt6 is isn't possible to update iterators by adding "+ 1" or "-1"
to them. In all versions moving outside of the map produces undefined
results, so the current "begin() - 1" code is not guaranteed to work.

Clean up the NextPrevWidgetFocus code a little bit by finding the
currently focused widget and starting the loop there.  If looping to
find the next widget, only run the second pass until getting back to
the current widget, instead of running to the end.