Simplify s_fileOpenCallbacks access in mythiowrapper.cpp.
The only two functions that access this hash are protected by a lock
so the queue can't change out from underneath them, so they can be
simple runs over the hash. The "register" function doesn't need to use
a QMutableHashIterator (which doesn't exist in Qt6) because it returns
immediately after modifying the hash.