qt6: Handle changes to the QSqlQuery class.
Qt6 removed the QSqlQuery::bindValues function to bind all query
values in one call. Working around this is easy since there is still a
function to bind a single value by name. Simply add a loop over the
binding values map to bind each value individually.

Qt6 also changed how the bindings are returned from a call to the
QSqlQuery::boundValues function. The return value is no longer a QMap
of named bindings and their values, but it is a QVariantList of the
binding values in numerical binding order.

There are two places where the binding values must be re-substituted
back into the query string. 1) If there is an error, rebinding the
values in order to resend the query is simple. They're already in the
proper order for numerical binding, so use a loop and rebind them
numerically. 2) For logging its a little more complicated. The
original query string has to be scanned multiple times to find/replace
each of the binding names with the correct value before the entire
query can be printed to the log.