ATSC Closed Captions
In avformatdecoder.cppp the closed captions packets are extracted from
the video stream and sent to the decoders for processing.
In commit 4880fe2427c1c83b861c3158fe7899865093cba0 of Nov 19, 2016  a check
is introduced to prevent potential out-of-bound memory access.
This check is not correct, causing the last closed caption packet
in each buffer to be always discarded. This causes all sorts of
issues in the rendering of the closed captions.
The check on out-of-bound memory access is now corrected.

Refs #326