Add fields to BlurayMetadata::ToMap.
There are two fields in this map whose value is presented as a single
UTF16 character.  New fields have been added to present the values in
a different way. They have had "<id>_str" fields added to the map that
presents the number as a string.  I.E. The titles field might be
presented as the two character string "12" instead of a one character
string containing QChar(12).