Be consistent about the format of m_origRecPriorityData.
This map is used to prevent saving all channel priority values after
one is changed. Unfortunately it was storing the original priorities
in the map as a single character, and then comparing them to the
string version of the priorities.  I.E. Stored in the map as QChar(1)
and then compared to "1", which is QChar(33). This meant that all
priorities were updated every time.

Change this to consistently use the string version of the priority
when referencing this map. Myth should now only update changed
priorities.