qt6: Fix parsing when reading from the command socket.
After converting the byte array to a string, qt5 ignored the NUL
character at the end when it was asked to trim white space from the
string. Qt6 does care about the NUL (and doesn't consider it white
space) so nothing is trimmed.  The QString::toInt function fails if it
sees any character other than a digit, thus it would always fail on
Qt6 where it had worked on Qt5.  Removing the padding byte at the end
of the array allows the same code to work on both version of Qt.