DVB Logical Channel Number offset per video source
With a mythtv-setup channel scan, the channel numbers assigned to the channels are the
DVB Logical Channel Numbers defined by the provider when they are present in the stream.
This are user-friendly numbers, usually in the range 1 to 1000 or 100 to 1000.

When you want to receive in MythTV more than one video source with DVB logical channel numbers
then the channel number ranges do usually overlap; this gives problems with duplicate channel
numbers and also channels from different video sources are interspersed in the program guide.

With the capability to add an offset for the logical channel numbers to each video source
it is possible to make the channel list non-overlapping.

The database schema is now version 1368.

Refs #345