qt6: Finish the DownloadMap/ArtworkMap QMap/QMultimap split.
In Qt6, the QMultiMap class is no longer based on the QMap class, so
you can no longer use the two almost interchangeably. Finish changing
variable declarations to the correct type, and in one case it is
necessary to copy the data from a QMap into a new QMultiMap.

Add to e912648ee8 LOCAL: Hack the DownloadMap/ArtworkMap problem so compiles finish.