Fix a couple of "potential null pointer dereference" warnings in mythmainwindow.cpp.
These warnings occur when building a harded image, so the optimization
is different from a normal build.

There are a number of places that call a function from the KeyContext
class using the form m_keyContexts.value(Context)->fn(). This looks
fine, except that the default constructor for this hash returns
nullptr. If "Context" doesn't already exist in the hash, this code
will crash.