Reversing list order for identical subtitles
After recording the first episode of the new season of "The Moodys", I found that the subtitle name "Episode 1" appears in both season 1 and season 2. Both ended up in the generated XML, starting with season 1. Since MythTV grabs the first item from the list, it was disappointing to see it pick the wrong one. To fix this, the code for tvmaze.py has been updated to list matching episode names, starting with the most recent release.