TVMaze: Improved search for subtitle
This has a number of enhancements to handle situations where there are
multiple matches on subtitle, and to better search for a subtitle
match when there are multiple shows with a matching title
(e.g. "The Flash").

I found some weaknesses in the string matching algorithm, so I modified
it. The Levenshtein distance is the minimum number of single character
edits required to change one string into another.
It's valuable information in trying to find a best match.
However, it's problematic to make decisions solely upon this distance.
If the strings are really long or really short, a simple rule, such as
"accept any distance < 6" can provide poor results. If we're comparing
"Risk" and "Call", the Levenshtein distance is 4, but it's a terrible
match. If we're comparing "In Sickness and in Health" and
"Sickness and Health", the Levenshtein distance value is 6. This would
fail the given rule, even though the strings are really a pretty good
match. 60% of the characters are in common. To establish how good of a
match we have, we'll use the Levenshtein distance to figure out the
ratio of matching characters in the longer string divided by the length
of the longer string. I call this 'match_quality'. This ratio ranges
from 0 (no match at all) to 1.0 (a perfect match).
This performs well for short, average, and long strings.

Refs #336

Signed-off-by: Roland Ernst <rcrernst@gmail.com>