Added tvmaze.py -N <title> <date time> support
For manual recordings, the program mythmetadatalookup starts its search
with a call to <grabber> -l <lang> -a <country> <title> <date time>,
where the <date time> is the local time stamp in ISO format.
For example:
tvmaze.py -l en -a US -N "CBS Sunday Morning" "2021-03-21 08:00:00"

This used to always fail, because grabbers like tvmaze.py were only
looking for a text <subtitle> as the last argument. I've added code
to check whether that field is in the form of a <date time>. If so,
this field is parsed to create a datetime which gets processed into
an argument for a call to tvmaze.get_show_episodes_by_date(), which
retrieves episode specific information from the database.

Updated and modified the --doctest handling to only be verbose when
the '--debug' option is also provided on the command line.

Refs #336

Signed-off-by: Roland Ernst <rcrernst@gmail.com>