clazy: Combine multiple calls to QString::arg in libmythtv/jobqueue.cpp.
Each call to QString::arg causes the Qt libraries to scan the string
to find the lowest numbered place marker, and it causes a memory
allocate to store the resulting integrated string. Consecutive calls
to ::arg with string arguments can be combined into a single call,
thus reducing the number of required string scans and (more
importantly) memory allocations.

This commit also cleans up a number of log message detail strings that
are originally created as a QString, are converted to a QByteArray,
are converted to a 'const char *', and are then used as the parameter
to QString::arg. Drop all the unnecessary conversions and pass the
original details QString in the call to ::arg.