clazy: Combine multiple calls to QString::arg in mythfrontend.
Each call to QString::arg causes the Qt libraries to scan the string
to find the lowest numbered place marker, and it causes a memory
allocate to store the resulting integrated string. Consecutive calls
to ::arg with string arguments can be combined into a single call,
thus reducing the number of required string scans and (more
importantly) memory allocations.