qt6: Overhaul the MythTVMenuNodeTuple structure (part 1)
Some of the code using this data structure won't compile in qt6
because of an error that the default definition for the operator=
function would be ill-formed. This is a problem when serializing and
deserializing this data structure.

Part 1 converts the structure to use an enum to specify which menu in
the TV data structure is being referenced, instead of including the
complete xml menu definition as part of the data structure.

Work over Pass around a menu id instead of a serialized MythTVMenu structure.