qt6: Overhaul the MythTVMenuNodeTuple structure (part 2)
Some of the code using this data structure won't compile in qt6
because of an error that the default definition for the operator=
function would be ill-formed. This is a problem when serializing and
deserializing this data structure.

Part 2 converts the structure to use an node path name in the
specified TV data structure member, instead of including the complete
xml text for the node.