Silence a couple of clang-tidy warnings. (libmythtv)
There are a couple of structures that seem to have changed
significantly between Qt5 and Qt6. When clang-tidy is run against a
Qt5 based build of MythTV it recommends one action, and when run
against a Qt6 build it recommends a contrary action. (Whether to move
or copy an object when it is passed a parameter to a function.)
Silence the warning that occurs in Qt6 until that becomes the
predominant version used in releases.