Update file size before skip forward
Get the actual file size before determining if it is
possible to skip the requested amount of seconds forward.
When the skip forward would go beyond the end of the file
the skip is modified so that it goes to the end of the
file instead. This gives the wrong result when the file
size information is not correct.

Backported from commit dbe165f81aa118871958b0663934a2f42edf464a in master.

Refs #352