Clear local MusicStreamListModified setting only when needed
As part of the "Running HouseKeeper." tasks there is every minute a check if the
RadioStreamUpdate task needs to run by checking the MusicStreamListModified setting.
As documented in the comment line:
// we are only interested in the global setting so remove any local host setting just in case
the local host value for this setting is cleared every time by writing to the database.
It should not be necessary to clear it more than once; it is now done only when
the value is actually not empty. For reading a setting a cache mechanism exists
so this fix does remove one database query every minute.
However, the main reason for this fix is that clearing the setting requires a write
lock on database table settings. When this table is being backed-up clearing the
setting is blocked until the table is available again.
During that time the Sat>IP recordings, and possibly also other type of recordings,
are also blocked. Blocking does not occur when the setting is read from the
database and also not when the setting is read from the settings cache.